feat: split functions into helper
